Add a way to make a Docker image IMUNES-ready
Usage:
$ ./custom.sh some/image:sometag [new_tag] [use_os]

For example:
$ ./custom.sh httpd web
will automatically figure out that httpd image is built on top of Debian
OS and will use custom/debian/Dockerfile to prepare the new image. The
new image will be tagged as imunes/template:web. If new_tag is not
provided, it will be tagged as imunes/template:custom.

This example won't automatically search for its base OS, but will use
custom/debian-12/Dockerfile (if available):
$ ./custom.sh jenkins/jenkins:lts-jdk17 jenkins debian-12

For now, only alpine and debian base OS's are available.